Advance-Traffic-Analysis/pyproject.toml
[project]
name = "advance-traffic-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "geopy>=2.4.1",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
